convenient conversion of tests
